Fix va_arg bug/crash on some 64-bit patforms
Calls to rb_scan_kw_args terminated the variable arg list with 0, but
that caused errors on the platform described below.

The problem appears to be that 0 gets passed via va_args as a type with
a different (smaller) size from "char *", which is the type used to
extract the arguments with va_arg.  This caused the va_arg processing
loop to run past the end.

The solution is to be sure to cast the final 0 argument in
rb_scan_kw_args calls as (char *).

Here is some info describing the patform where this problem was
discovered and fixed...

$ uname -a
Linux paper1 2.6.31-20-generic #58-Ubuntu SMP Fri Mar 12 04:38:19 UTC 2010 x86_64 GNU/Linux

$ gcc --version
gcc (Ubuntu 4.4.1-4ubuntu9) 4.4.1
Copyright (C) 2009 Free Software Foundation, Inc.
This is free software; see the source for copying conditions.  There is NO
warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.