Have Rakefile build ext/rb_pgplot.c for gem file
Now that ext/rb_pgplot.c is included in the gemspec as an "extra_rdoc_file", it
is necessary to include it in the gem file or "rake gem" complains.  Since this
file generation does not need any runtime specific information, it can be
generated when the gem file is being built.  This commit automates the
generation of this file when using "rake gem".

It also changes the rdoc --exclude options in the gemspec so that the
rb_pgplot.c file is not excluded (as happened using gem version 1.8.23 and rdoc
version 3.9.4).